+++ src/ambition/version_details.cc
@@ -27,6 +27,7 @@ std::vector<VersionUpdates> getVersionDetails(
           _("Fixed diagonal lines (unit paths, waypoints, rally points, building links) looking stretched out compared to orthogonal lines."),
           _("Fixed an issue where accepting a friendly treaty would override a declaration of war."),
           _("Preserve all the user options (Report Backgrounds, News Display, Unit Identity Display, and Unit Path Display) when loading a save file or starting a scenario (which loads a save file, internally)."),
+          _("Fixed ship washes (the waves around, and the wake behind, ships) appearing only when hovering over ships."),
           _("Avoid potential issues from loading newer incompatible save files that may be saved by future versions with more features."),
         },
       },